gradio ompatiblity
the .style() methods are no longer supported. The style attributes need to be passed in the preceding method.

The 'source' attribute for gr.Image, etc. has been replaced by 'sources' (plural).

Note that the requirement.txt file does not specify a version for gradio and so the code needed to be updated as it was no longer working.